+++ jsonlink/src/main/java/io/freedriver/jsonlink/config/v2/AnalogAlertCondition.java
@@ -3,16 +3,16 @@ package io.freedriver.jsonlink.config.v2;
 public enum AnalogAlertCondition {
     LESS_THAN {
         @Override
-        public boolean test(float benchmark, float actual) {
+        public boolean test(double benchmark, double actual) {
             return actual < benchmark;
         }
     }, // Percent
     MORE_THAN {
         @Override
-        public boolean test(float benchmark, float actual) {
+        public boolean test(double benchmark, double actual) {
             return actual > benchmark;
         }
     }; // Percent
 
-    public abstract boolean test(float benchmark, float actual);
+    public abstract boolean test(double benchmark, double actual);
 }
